requirements.txt
discord.py
langchain==0.1.0
openai
chromadb
tiktoken